requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-SocketIO>=5.3.6
python-socketio>=5.8.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.31.0
google-generativeai>=0.8.3
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
gunicorn>=21.0.0
eventlet>=0.33.3
# FASE 1: Mejoras críticas
marshmallow>=3.20.0
flask-marshmallow>=0.15.0
Flask-Limiter>=3.5.0
Flask-Migrate>=4.0.5
loguru>=0.7.2
# FASE 2: Performance
Flask-Caching>=2.1.0
redis>=5.0.0
Flask-Compress>=1.14
tenacity>=8.2.3
# FASE 3: Estadísticas y exportación
pandas>=2.0.0
openpyxl>=3.1.0
